migrate list view to omfp declarative dashboard
