+++ projects/wc/src/app/components/generic-ui/list-view/resource-table-card/resource-table-card.component.html
@@ -1,8 +1,8 @@
 <mfp-declarative-table-card
   [resources]="resources()"
   [config]="config()"
-  (tableRowClicked)="tableRowClicked.emit($event)"
-  (loadMoreResources)="loadMoreResources.emit()"
-  (paginationLimitChanged)="paginationLimitChanged.emit($event)"
-  (actionButtonClick)="buttonClick.emit($event)"
+  (tableRowClicked)="navigateToResource($event)"
+  (loadMoreResources)="loadMore()"
+  (paginationLimitChanged)="onLimitChange($event)"
+  (actionButtonClick)="executeAction($event)"
 />
