+++ projects/wc/src/app/components/generic-ui/list-view/resource-table-card/resource-table-card.component.ts
@@ -1,31 +1,251 @@
 import {
   ChangeDetectionStrategy,
   Component,
+  DestroyRef,
   ViewEncapsulation,
-  output,
+  computed,
+  effect,
+  inject,
   input,
+  signal,
 } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { DeclarativeTableCard, TableCardConfig, ValueCellButtonClickEvent } from '@openmfp/ngx';
+import { LuigiClient } from '@luigi-project/client/luigi-element';
 import {
-  DeclarativeTableCard,
-  TableCardConfig,
-  ValueCellButtonClickEvent,
-} from '@openmfp/ngx';
-import { Resource } from '@platform-mesh/portal-ui-lib/models';
+  Resource,
+  ResourceListResult,
+  ResourceSubscriptionResult,
+} from '@platform-mesh/portal-ui-lib/models';
+import {
+  ErrorHandlerService,
+  ResourceNodeContext,
+  ResourceService,
+} from '@platform-mesh/portal-ui-lib/services';
+import {
+  buildResourcePath,
+  generateGraphQLFields,
+  isNamespacedResource,
+  mergeListWithSubscriptionResult,
+} from '@platform-mesh/portal-ui-lib/utils';
+import { finalize } from 'rxjs/operators';
+import { executeButtonAction } from '../../../../utils/field-definition.utils';
+import { addSearchParams } from '../../../../utils/set-search-params';
 
 @Component({
   selector: 'pm-resource-table-card',
   standalone: true,
   imports: [DeclarativeTableCard],
   templateUrl: './resource-table-card.component.html',
-  encapsulation: ViewEncapsulation.None,
+  encapsulation: ViewEncapsulation.Emulated,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ResourceTableCard {
-  resources = input.required<Resource[]>();
-  config = input.required<TableCardConfig>();
+  private resourceService = inject(ResourceService);
+  private errorHandlerService = inject(ErrorHandlerService);
+  private destroyRef = inject(DestroyRef);
 
-  readonly tableRowClicked = output<Resource>();
-  readonly loadMoreResources = output<void>();
-  readonly paginationLimitChanged = output<number>();
-  readonly buttonClick = output<ValueCellButtonClickEvent<Resource>>();
+  LuigiClient = input.required<LuigiClient>();
+  context = input.required<ResourceNodeContext>();
+
+  resources = signal<Resource[]>([]);
+  resourceDefinition = computed(() => this.context().resourceDefinition);
+  columns = computed(() => {
+    let columns = this.resourceDefinition()?.ui?.listView?.fields ?? [];
+
+    const readyCondition = this.resourceDefinition()?.readyCondition;
+    if (readyCondition) {
+      columns = [
+        {
+          ...readyCondition,
+          uiSettings: {
+            ...readyCondition.uiSettings,
+            displayAs: 'alert',
+            columnWidth: '30px',
+          },
+        },
+        ...columns,
+      ];
+    }
+
+    return columns;
+  });
+
+  totalItemsCount = computed(
+    () => this.resources().length + this.remainingItemCount(),
+  );
+  paginationLimit = signal<number>(5);
+  remainingItemCount = signal<number>(0);
+  hasMore = signal<boolean>(false);
+  resourceVersion = signal<string | undefined>(undefined);
+
+  config = computed<TableCardConfig>(() => ({
+    tableConfig: {
+      fields: this.columns(),
+      totalItemsCount: this.totalItemsCount(),
+      paginationLimit: this.paginationLimit(),
+      hasMore: this.hasMore(),
+    },
+  }));
+
+  private isNamespaced = computed(() => isNamespacedResource(this.context()));
+  private currentContinueToken: string | undefined = undefined;
+  private isLoadingList = false;
+
+  constructor() {
+    effect(() => {
+      this.currentContinueToken = undefined;
+      this.list(true);
+    });
+
+    effect((onCleanup) => {
+      const version = this.resourceVersion();
+      if (!version) return;
+      const sub = this.subscribeToResourceChange(version);
+      onCleanup(() => sub.unsubscribe());
+    });
+  }
+
+  private subscribeToResourceChange(version: string) {
+    const fields = this.getListQueryFields();
+    const resourceDefinition = this.context().resourceDefinition!;
+    const queryOperation = buildResourcePath({
+      apiGroup: resourceDefinition.apiGroup,
+      version: resourceDefinition.version,
+      entity: resourceDefinition.entityCollection,
+    }) as string;
+
+    return this.resourceService
+      .resourceChangeSubscription(
+        queryOperation,
+        fields,
+        this.context(),
+        version,
+        false,
+      )
+      .subscribe({
+        next: (value) => {
+          if (!value) return;
+          this.mergeResourcesWithSubscriptionResult(value);
+        },
+        error: (_error) => {
+          this.LuigiClient().uxManager().showAlert({
+            text: 'Error while updating list with new data. To see new updates, refresh the page.',
+            type: 'error',
+          });
+        },
+      });
+  }
+
+  private resetPagination() {
+    this.currentContinueToken = undefined;
+    this.resources.update((v) => v.slice(0, this.paginationLimit()));
+    this.hasMore.set(this.resources().length < this.totalItemsCount());
+  }
+
+  onLimitChange(limit: number) {
+    this.paginationLimit.set(limit);
+    this.resetPagination();
+  }
+
+  loadMore() {
+    if (!this.hasMore()) return;
+    this.list();
+  }
+
+  list(isInitialLoad: boolean = false) {
+    if (this.isLoadingList) return;
+    this.isLoadingList = true;
+
+    const fields = this.getListQueryFields();
+    const resourceDefinition = this.getResourceDefinition();
+    const queryOperation = buildResourcePath({
+      apiGroup: resourceDefinition.apiGroup,
+      version: resourceDefinition.version,
+      entity: resourceDefinition.entityCollection,
+    }) as string;
+
+    this.resourceService
+      .list(queryOperation, fields, this.context(), false, {
+        limit: this.paginationLimit(),
+        continue: this.currentContinueToken,
+      })
+      .pipe(
+        finalize(() => (this.isLoadingList = false)),
+        takeUntilDestroyed(this.destroyRef),
+      )
+      .subscribe({
+        next: (result: ResourceListResult) => {
+          if (isInitialLoad) {
+            this.resources.set(result.items ?? []);
+          } else {
+            this.resources.update((values) => {
+              const map = new Map(values.map((i) => [i.metadata.name, i]));
+              (result.items ?? []).forEach((i) => map.set(i.metadata.name, i));
+              return [...map.values()];
+            });
+          }
+          this.resourceVersion.set(result.resourceVersion);
+          this.hasMore.set(!!result.continue);
+          this.currentContinueToken = result.continue;
+          this.remainingItemCount.set(result.remainingItemCount || 0);
+        },
+        error: (error) => {
+          this.errorHandlerService.handleError(error);
+        },
+      });
+  }
+
+  private mergeResourcesWithSubscriptionResult(
+    subscriptionResult: ResourceSubscriptionResult,
+  ) {
+    this.resources.set(
+      mergeListWithSubscriptionResult(this.resources(), subscriptionResult, {
+        getItemKey: (item) => item.metadata?.name,
+        mapSubscriptionObjectToItem: (object) => object,
+      }),
+    );
+  }
+
+  navigateToResource(resource: Resource) {
+    const resourceDefinition = this.context().resourceDefinition;
+    if (!resourceDefinition?.ui?.detailView) return;
+
+    if (!resource.metadata.name) {
+      this.LuigiClient().uxManager().showAlert({
+        text: 'Resource name is not defined',
+        type: 'error',
+      });
+      throw new Error('Resource name is not defined');
+    }
+
+    addSearchParams({
+      namespace: this.isNamespaced() ? resource.metadata.namespace : undefined,
+    });
+    this.LuigiClient().linkManager().navigate(resource.metadata.name);
+  }
+
+  executeAction(event: ValueCellButtonClickEvent<Resource>) {
+    executeButtonAction(this.LuigiClient(), event.field, event.resource);
+  }
+
+  private getListQueryFields() {
+    const additionalFields = [{ property: 'metadata.deletionTimestamp' }];
+    if (this.isNamespaced()) {
+      additionalFields.push({ property: 'metadata.namespace' });
+    }
+    return generateGraphQLFields(this.columns().concat(additionalFields));
+  }
+
+  private getResourceDefinition() {
+    const resourceDefinition = this.resourceDefinition();
+    if (!resourceDefinition) {
+      this.LuigiClient().uxManager().showAlert({
+        text: 'Resource definition is not defined',
+        type: 'error',
+      });
+      throw new Error('Resource definition is not defined');
+    }
+    return resourceDefinition;
+  }
 }
